tafsir

+++ index.js
@@ -4,6 +4,7 @@ const fs = require('fs')
 
 const chapterPath = 'https://web-api.qurankemenag.net/quran-surah'
 const versePath = 'https://web-api.qurankemenag.net/quran-ayah'
+const tafsirPath = 'https://web-api.qurankemenag.net/quran-tafsir'
 
 async function generateChapters() {
   const res = await axios.get(chapterPath, {
@@ -34,7 +35,7 @@ async function generateVerses() {
       }
     })
 
-    if (chapter === 1) {
+    if (chapter === 1 && fs.existsSync('output/verses.csv')) {
       await fs.promises.rm('output/verses.csv')
     }
 
@@ -51,5 +52,35 @@ async function generateVerses() {
   }
 }
 
+async function generateTafsir() {
+  const verses = Array.from({ length: 6236 }, (_, i) => i + 1)
+
+  for (const verse of verses) {
+    const res = await axios.get(`${tafsirPath}/${verse}`, {
+      headers: {
+        'Origin': 'https://quran.kemenag.go.id'
+      },
+    })
+
+    if (verse === 1 && fs.existsSync('output/tafsir.csv')) {
+      await fs.promises.rm('output/tafsir.csv')
+    }
+
+    const csv = await json2csv(res.data.data, {
+      prependHeader: verse === 1,
+      keys: ['id', 'surah_id', 'ayah', 'tafsir.wajiz']
+    })
+
+    if (!fs.existsSync('output')) {
+      await fs.promises.mkdir('output')
+    }
+
+    await fs.promises.appendFile('output/tafsir.csv', csv + '\n')
+
+    console.log(`verse ${verse} tafsir processed`)
+  }
+}
+
 generateChapters()
 generateVerses()
+generateTafsir()
